+++ src/entity/GameObject.cpp
@@ -12,11 +12,9 @@
 
 #include "GameObject.h"
 
-
-
 int GameObject::last_id = 0;
 
-GameObject::GameObject(const std::string &name, Point3d const &pos) :
+GameObject::GameObject(const std::string &name, Point3f const &pos) :
         id(++last_id), name(name), position(pos)
 {
     //id = ++last_id;
